+++ packages/blackborne/src/components/Breadcrumbs/index.ts
@@ -1,2 +1,4 @@
-export { Breadcrumb, Breadcrumbs } from './Breadcrumbs';
-export type { BreadcrumbProps, BreadcrumbsProps } from './Breadcrumbs';
+export { Breadcrumb } from './Breadcrumb';
+export type { BreadcrumbProps } from './Breadcrumb';
+export { Breadcrumbs } from './Breadcrumbs';
+export type { BreadcrumbsProps } from './Breadcrumbs';